requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
groq==0.4.2
pydantic==2.4.2
python-dotenv==1.0.0

# Vector DB and embeddings - using FAISS
sentence-transformers==2.2.2
faiss-cpu==1.7.4
numpy==1.23.5
scikit-learn==1.3.2

# Data processing
pandas==2.0.3
tiktoken==0.5.1

# NLTK and transformers
transformers==4.35.0
nltk==3.8.1
torch==2.1.0

# Testing
pytest==7.4.3
httpx==0.25.1

# Utilities
typing-extensions==4.8.0